fix: bug receiving props inside patient list item
